moved storelist functions to home
