+++ src/components/Home/actionTypes.js
@@ -1,3 +1,5 @@
 export const SET_SEND_WITH_PHOTO = 'SET_SEND_WITH_PHOTO';
 export const SET_USERS = 'SET_USERS';
 export const SET_USER_REFERENCE = 'SET_USER_REFERENCE';
+export const SET_STORELIST = 'SET_STORELIST';
+export const SET_LOAD_STORE_LIST_ERROR = 'SET_LOAD_STORE_LIST_ERROR';
