+++ src/components/Home/actions.js
@@ -1,7 +1,10 @@
+import fetchItems from '../../utils/honestyStore.js';
 import {
   SET_SEND_WITH_PHOTO,
   SET_USERS,
-  SET_USER_REFERENCE
+  SET_USER_REFERENCE,
+  SET_STORELIST,
+  SET_LOAD_STORE_LIST_ERROR
 } from './actionTypes';
 import initFirebase from '../../utils/firebase';
 import firebase from 'firebase/app';
@@ -27,6 +30,26 @@ function setUsers(users) {
     users
   };
 }
+
+export function setStoreList(storeList) {
+  return {
+    type: SET_STORELIST,
+    storeList
+  };
+}
+
+export function setLoadStoreListError(loadStoreListError) {
+  return {
+    type: SET_LOAD_STORE_LIST_ERROR,
+    loadStoreListError
+  };
+}
+
+export const loadStoreList = () => async dispatch => {
+  return await fetchItems().then(storeList =>
+    dispatch(setStoreList(storeList))
+  );
+};
 
 export const attemptLoadUsers = async () => {
   initFirebase();
