--- src/components/StoreList/actionTypes.js
@@ -1,2 +1,0 @@
-export const SET_STORELIST = 'SET_STORELIST';
-export const SET_LOAD_STORE_LIST_ERROR = 'SET_LOAD_STORE_LIST_ERROR';
