--- src/components/StoreList/actions.js
@@ -1,22 +1,0 @@
-import {SET_STORELIST, SET_LOAD_STORE_LIST_ERROR} from './actionTypes';
-import fetchItems from '../../utils/honestyStore.js';
-
-export function setStoreList(storeList) {
-  return {
-    type: SET_STORELIST,
-    storeList
-  };
-}
-
-export function setLoadStoreListError(loadStoreListError) {
-  return {
-    type: SET_LOAD_STORE_LIST_ERROR,
-    loadStoreListError
-  };
-}
-
-export const loadStoreList = () => async dispatch => {
-  return await fetchItems().then(storeList =>
-    dispatch(setStoreList(storeList))
-  );
-};
